Updates to srcSlice. Directing the current policy to generate profiles
